test: contract tests also green on real parser (round-trip)
- doc_via_real() = serialize(mock) -> parse, so every contract test now runs
  against both the mock and the real parser/writer
- writer now reuses an existing matching import/export directory instead of
  always appending a .peimp/.peexp section -> serialization is idempotent
- fixed import renderer dropping DLL name strings
- roundtrip_test: parse(serialize(doc)) preserves headers/sections/imports/
  exports, double round-trip stable, malformed input rejected
- 32 tests green, clippy clean

+++ crates/pe-rs/src/io/pe/export_render.rs
@@ -31,7 +31,7 @@ pub fn render_export_table(exports: &ExportTable, base_rva: Rva) -> Result<Rende
     syms.sort_by_key(|s| s.ordinal);
     let named: Vec<&ExportSymbol> = syms.iter().filter(|s| s.name.is_some()).cloned().collect();
 
-    let n_funcs = syms.len() as usize;
+    let n_funcs = syms.len();
     let funcs_off = 40usize;
     let names_off = funcs_off + n_funcs * 4;
     let ordinals_off = names_off + named.len() * 4;
